+++ implementation/js/src/infrastructure/lib+-/marionette/region-manager.js
@@ -3,11 +3,36 @@
  *
  * @author Tim Lauv
  * @created 2016.02.05
+ * @updated 2017.03.10
  */
 
 ;(function(app){
 
 	_.extend(Marionette.RegionManager.prototype, {
+
+		// Add an individual region to the region manager,
+		// and return the region instance
+		// +We give the region object its own name as the _name property.
+		addRegion: function(name, definition) {
+		    var region;
+
+		    var isObject = _.isObject(definition);
+		    var isString = _.isString(definition);
+		    var hasSelector = !!definition.selector;
+
+		    if (isString || (isObject && hasSelector)) {
+		        region = Marionette.Region.buildRegion(definition, Marionette.Region);
+		    } else if (_.isFunction(definition)) {
+		        region = Marionette.Region.buildRegion(definition, Marionette.Region);
+		    } else {
+		        region = definition;
+		    }
+
+		    region._name = name;
+		    this._store(name, region);
+		    this.triggerMethod("region:add", name, region);
+		    return region;
+		},
 
 	    // Close all regions in the region manager, but
 	    // leave them attached