+++ implementation/js/src/infrastructure/lib+-/marionette/view.js
@@ -849,7 +849,7 @@
 			if(!this.data && !this.useParentData){
 				if(this.parentRegion)
 				    this.parentRegion.once('show', function(){
-				    	//this is to make sure local data ready always fires after navigation-chain completes (e.g after view:navigate-to)
+				    	//this is to make sure local data ready always fires after region/view animation completes.
 				    	this.currentView.triggerMethodInversed('ready');
 				    	//note that form view will not re-render on .set(data) so there should be no 2x view `ready` triggered.
 				    });
@@ -875,9 +875,6 @@
 		        this.data = tmp;
 		    }
 		    if (this.data){
-		    	//mark local data case, so first data ready can be fired after navigate-to (after region:show)
-				if(_.isPlainObject(this.data))
-					this._delayFirstTimeLocalDataReady = true;
 		        this.set(this.data);
 		    }
 		});